fix: change potoken verify strategy to avoid bot detection
- fix the logic of the reload function

Bot detection should happen less often now. And if it does get detected as a bot, a new token will be generated and it should guarantee monty to not be detected as a bot on the next request. Other than that the reload function can now work properly. Whenever there's a new update the owner can just <prefix>reload to reload everything and the changes will be applied on the fly without having to restart main.py as long as the changes happen in the cogs, not in main.py