fix, feat: implement music class
Saves a music into an instance of a dedicated class with its own methods. Additionally, implement a functionality to check the expiry of the music's source before playing it. This will fix the bot spamming the text channel when the source gets expired. Also fixes the requirements.txt to be UTF-8 instead of whatever gibberish it became because of windows